third_party/WebKit/LayoutTests/http/tests/payments/payment-app-manager.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../../serviceworker/resources/test-helpers.js"></script>
<script>

promise_test(test => {
    var registration;
    var script_url = 'resources/empty-worker.js';
    var scope = 'resources/';

    return service_worker_unregister_and_register(test, script_url, scope)
      .then(r => {
          registration = r;
          return wait_for_state(test, registration.installing, 'activated');
        })
      .then(state => {
          assert_equals(state, 'activated');
          return registration.paymentAppManager.setManifest({
              label: 'Payment App'
            });
        })
      .then(result => {
          assert_equals(result, undefined);
        })
      .catch(unreached_rejection(test));
  }, 'setManifest() should resolve with undefined if '
      + 'setManifest() is succeeded.');

promise_test(test => {
    var registration;
    var script_url = 'resources/empty-worker.js';
    var scope = 'resources/';

    return service_worker_unregister_and_register(test, script_url, scope)
      .then(r => {
          registration = r;
          return wait_for_state(test, registration.installing, 'installed');
        })
      .then(state => {
          assert_equals(state, 'installed');
          return registration.paymentAppManager.setManifest({
              label: 'Payment App'
            });
        })
      .then(result => {
          assert_equals(result, undefined);
        })
      .catch(unreached_rejection(test));
  }, 'If registration has waiting worker, then wait for active worker and '
      + 'then setManifest() is succeeded.');

promise_test(test => {
    var registration;
    var script_url = 'resources/empty-worker.js';
    var scope = 'resources/';

    return service_worker_unregister_and_register(test, script_url, scope)
      .then(r => {
          registration = r;
          registration.unregister();
          return wait_for_state(test, registration.installing, 'redundant');
        })
      .then(state => {
          assert_equals(state, 'redundant');
          assert_equals(registration.installing, null);
          assert_equals(registration.waiting, null);
          assert_equals(registration.active, null);
          return registration.paymentAppManager.setManifest({
              label: 'Payment App'
            });
        })
      .then(unreached_fulfillment(test))
      .catch(error => {
          assert_equals(error.name, 'InvalidStateError');
        });
  }, 'If registration has no active worker, no waiting worker and '
      + 'no installing worker, then throws InvalidStateError.');

</script>
